Fast matcher: Return the match-list as an array instead of a linked list
This has less overhead due to:
1. It is faster just to fill-in array elements.
2. Memory allocation:
 - Can just malloc() all the array elements at once, instead of an
   malloc() per list element.
 - No need to free() until finishing handling the sentence.

However, in case of returning an empty list it has (very) slight more
overhead because it then returns one element with a NULL disjunct
instead of just NULL (can be fixed with a few more code lines, but the
possible gain is very small).

This makes corpus-fixes.batch 5% faster, corpus-basic.batch somewhat
less than 1% faster, but has a negligible effect of Russian and long
English sentences (as tested with the "In vivo studies ..." sentence).

Note: Because currently the list is returned in reverse order, and in
this change the array is in the match-list table order (because it is
more efficient and elegant code this way), linkages of the same cost are
returned in a different order. This makes 2 tests to fail (to be
fixed in the next commit).