Bug fix: Use "’", not "`", as an additional contraction mark character
This fixes an error I introduced. It kept unnoticed up to now because
it somehow doesn't effect the very few example sentences that use this
character.

This fix also includes a code change, because my original function to
identify a contraction is not Unicode aware (so using "’", which is a
3-byte Unicode character, introduced strange bugs).

FIXME: Currently contraction handling is totally English-centric. This
should be fix too. See also issue #268.