Set word_in_dict if there is any suffix split alternative
This bug causes extra parses for sentences, in which a word is not in
the dictionary, there is more than one alternative to split it, and the
last alternative has a non-stem word in the position of the stem.  The
actual occurrence of the problem depends on the suffixes order in the
SUF list - which determines the order of these alternatives.

The bug happens because word_in_dict is **assigned** the success value
of add_alternative_with_subscr() - i.e. whether it finds the word that
is in the stem position in the dict as stem (as opposed to a regular
word). So it is overwritten to false if a last alternative generation
fails.

As a result, because the word is not in the dictionary and appears as
if it cannot split, a regex lookup is done, which adds extra parses.

This happens only rarely (I found one case in the Russian corpus file
when I tested the effect of sorting the SUF list). The fix is of course
to use OR operation.