Further fix the long-irritating backtic bug
Just now a single tic is unknown in the dict:
linkparser> They have told of the soldiers' fear
LEFT-WALL they have.v told.v-d of the soldiers.n '[?].n fear.n

We have now in both en/4.0.dict.m4 and en/4.0.dict:
"’" "''": YP- & (({AL-} & {@L+} & (D+ or DD+)) or [[<noun-main-x>]] or DP+);

It turns out that '' is not actually a good idea to generate a single
tic in m4:
linkparser> !!'
Token "'" matches nothing in the dictionary.
Token "'" matches nothing in the dictionary.
linkparser> !!''
Token "''" matches:
    ''                              29  disjuncts
Token "''" expressions:
    ''                       YP- & (({AL-} & {@L+} & (D+ or DD+)) ...

This is because it is not a documented way in m4 to generate a single
tic and it just happened to work by chance when first tried, due to a
"backtic imbalance" that existed at that time.

This change fixes the wrong m4 quote opening caused by "`til",
and reverts the use of tic doubling. As a result, the following is
fixed too:
data/en/4.0.dict       :% Give AN+ a cost, because in general, we dont want to conjoin nouns,
data/en/4.0.dict.m4    :% Give AN+ a cost, because in general, we don't want to conjoin nouns,

To sum up:
Any use of a backtic in the dict.m4 file needs to be protected by:
changequote(\,/)dnl
Here a ` is used ...
changequote dnl

and no special handling is needed for '.