Python binding: Return an empty iterator for no linkages
This is a fix for the failed try (commit message):
Python binding: Fix a StopIteration exception on no linkages
that still makes an iteration when there are no linkages, such as in:
linkages = sent.parse()
for linkage in linkages:
    ....
causing the program (e.g. example.py) to abort on a sentence with no
linkages.