Fix "Assertion (word_xnode != NULL) failed" in SAT solver
Repeat by:
echo TEST | link-parser -use-sat

Result:
Assertion (word_xnode != NULL) failed at .../link-grammar/sat-solver/word-tag.cpp:130: NULL next X_node for var w1d0

This problem happened in commit: "add_empty_word(): Invoke only when needed"
due to a wrong assumption in my original code.

** I don't have a prove that the fix is correct. **
However, I verified that exactly the same linkages are produced by the
regular parser and the SAT parser, as follows:

1. Checked corpus: en/corpus-basic.batch.
2. All parameters are the same as in the batch file, besides: -limit=100000.
3. Up to 50000 linkages are actually produced: -auto-next-linkage:50000
   (the default is 1024).
4. The tests script suppresses identical linkages, and sorts them
   according to the alphabetic value of their graphical representations.
   Then a diff is used to compare the results of the regular and the SAT
   parsers.

BTW, if the batch default -limit=10000 is used, the results of all the
sentences in the said corpus batch are the same, beside for the sentence
which starts with "Voting in favor", which is missing a few linkages
on the SAT batch run. But when enough linkages are produced (the
parameters in 2 and 3 above) these missing linkages appear too.

It will be a good idea to add yet another assert() to validate that
the used word_xnode is the correct one, but I till need to find out how
to do that.