_WIN32: Code cleanup for last-resort data directory
When the language directory is not found in ".", ".\data", ".." or
"..\data", the _WIN32 code uses as the last-resort data directory the
program's EXE file location. This is done in dictionary_get_data_dir().

The code for that is messy. It also has a section to support UTF-16 file
names, which actually supports only ASCII file names (as observed in the
FIXME there).

This fix is equivalent to the original code (still not supporting
non-ASCII path names), but is much clearer.

More in this fix:
- Use native Windows PathRemoveFileSpecA() instead of path_get_dirname()
  (which gets removed by this change).
- Use MAX_PATH_NAME (defined in structures.h) instead of MAX_PATH
  (from a Windows header file).
- Instead of printing messages on success, print error messages on
  failure.