Windows bug fix: afdict string handling
The QUOTES and BULLETS strings are converted by mbsrtowcs() using the
current locale, which is not a UTF-8 one under Windows. But 4.0.affix,
as the rest of our data files, is UTF-8 encoded. So the result of this
conversion is junk. The searched character is converted correctly, so
if it is a non-ASCII one it is never found.

This doesn't work correctly on Windows (not that the quotes are UTF-8
characters):
„What are you doing?“ she asked.

This fix keeps the internal representation as UTF-8 (instead of
wide char) so no conversions are needed.