New API function linkgrammar_get_dict_locale(Dictionary)
Return a locale for the given dictionary, in the OS format (suitable for
use in setlocale()).

If <dictionary-locale> defined in the given dict and it is valid,
return it. Valid values are In the format LL4cc+ or the string C+.
The later is for specifying the C locale, intended to be used for
romanized dictionaries (EN4us+ and others can be used too, but
the intention is to imply that spell checking is to be disabled.)

Else return the user locale in a hope it is appropriate for this dict.
On all OSs look at the appropriate LC_* environment variables. Their
value is assumed to be in the OS native format that uses language and
territory codes (not a number or long names). On Windows, if these
variables are not set, return the user default locale.
This feature eliminates the need to duplicate a lot of related library
code in the link-parser program.

For example:
<dictionary-locale>: HE4il+;
Returned value:
Windows: he-IL
Other OSs: he_IL.UTF-8