link-parser: Use fgets() on stdin file input
Currently link-parser, when compiled with editline, uses it also for
file input from stdin. This caused the history overwrite by input batch
files.

This fix uses fgets() to read input files from stdin. This also makes
the program slightly faster by eliminating the overhead of readline.
The history-rewrite fix has been removed in a recent commit
because the fix here makes it unnecessary.

This fix also removes the fgetws/wcstombs for Windows, that didn't work
as intended anyway: It was used only for !file read, due to the
condition (stdin == in) which is true also for batch files read by
stdin.  It also converted the input using the current code page, while
the batch files are in utf-8 and needs to be read without any
conversion.

Note that on Windows, only \n (and not \r) needs to be deleted since
!files uses fopen without "b", and stdin is in text mode by default.