Windows: Add support for console UTF-8 read/write
The added get_console_line() reads UTF-16 from the console and converts
it to UTF-8. The console code page is not involved in that.

The added win32_set_utf8_output() sets the output (also for files pipes)
to have no conversions, so the internal UTF-8 strings can be output as
they are.