Check for long input lines.
If readline is not used, a fixed input buffer of MAX_INPUT is used.
No indication is provided if the input line is longer.
This change adds a notification.

The processing of such a line still doesn't make much sense since it is
truncated. In addition, the next input line is the next part of the
too-long input line.

Possible solutions:
1. Support "infinite" long lines (but there's no point to do that).
2. Discard long lines (including their continuation part(s).