Strip whitespace from end of lines
It turned out it is needed after all, but now it is done in one place.
Also remove the check in special_command() of '\n' at line start - it's
surely not needed after this change.

Note that we cannot strip \xA0 here because this value can be the last
byte of a UTF-8 character (unless we make a more complicated check).