Add locale-handling functions
1. newlocale_LC_CTYPE(locate_string): Create a locale object from a
locale string.  It will be kept in the Dictionary structure, and will be
used by the isw*_l() character slasifying functions.

2. is_known_locale(locale_string): Validate that the given locale is
know to the system. This is needed in order not to put in the Dictionary
structure a malformed locale name, that causes a fault when it is
actually used.

3. set_utf8_program_locale(void): On Linux it is needed to set the
program locale to one with a UTF-8 codeset. To make things simpler, the
program locale is set if _WIN32 is not defined. The reason is that
Linux doesn't have mbrtowc_l()/mbsrtowcs_l() and mbrtowc() doesn't know
about UTF-8 when in the "C" locale.