Rename is_known_locale() to try_locale()
In case the system doesn't have locale_t, we cannot use newlocale() in
order to validate whether the given locale is valid, and must actually
try to validate that to set the locale by setlocale(). If the locale is
valid, this check has of course a side effect of setting the process
locale. However, this is fine, since we need to do it anyway.

Rename the function to describe better.