Test suite: Test using the "en" dict with user "tr" locale
The test sentence starts with "I". On 3.5.7 this test fails.
On this version the "en" dictionary defines its locale to be
en_US, and it is used instead.

A similar test can be added for using the "tr" dict with a non-"tr"
locale, but currently I don't have a test sentence that starts with "I"
(and also the "tr" dict needs to support parsing it).