Windows: Cast GetLastError() with int
It returns DWORD, which is defined as int or long in different
compilation models of MSVC and MinGW, but it is always 32 bits in any
case. The casts are needed to eliminate printf format warnings when it
happens to be defined as long.

A similar cast is done for LCID.