format_locale: Restore previous code - needed for MinGW
setlocale() on MinGW doesn't know about ll-CC locale conventions.
It needs the full names Language_Country convention. There are reports
on the Internet that some Windows versions also need the full locale
names. This function generates the full locale name from the short one.

Note that only two-component locales are supported for now. This is
enough for the current dictionaries. This can be fixed.