MINGW: Fix Unicode output
An ifdef _MSV_VER has not been replaced by _WIN32.
However, this code is needed for native Windows, independent of the
compiler version.
Fix it by using _WIN32 instead.
There are still ifdef's of _MSV_VER in the code, but they are all
(hopefully) for things that are specific for MSVC.

FIXME:
However, it after this fix it prints fine only when running under
Cygwin.  It turns out the problem is triggered by the stdio implementation
of MinGW that is used (in order to handle %z) instead of the native
Windows library stdio (still investigated).