MINGW: Define VISTA in a similar way
On MSVC it is impossible to set _WIN32_WINNT to NTDDI_VISTA and instead
there is a need to set it to _WIN32_WINNT_VISTA, even though they have
the same numerical value (this is enforced by a check in a header file).

So this is done here for MinGW too.