Allow detecting if spell guess is not configured
Currently the tests suite fails for libraries in which spell guess is
not configured.

This change allows to programmatically find if spell guess is enabled,
by setting the spell option and checking if it is not 0.

FIXME: The library should offer a better way to handle optional features,
like providing a "feature string".