Windows: More general version variable replacement script
The current script can only replace the atoconf LINK_.*_VERSION variables,
and the file name link-features.h.in is hard coded.

Replace its operation to be a filter, so it can process also the
__init.py.in files for Python. Also add the variable VERSION,
and enable it to replace $VERSION (in addition to @VERSION@).

The script is also renamed according to its new function.