SAT: Bug fix - preallocate linkages memory
Currently the linkage array in the SAT parser is allocated on demand -
for each new linkage it is enlarged by one element, using realloc().

The problem is that realloc() may return a pointer to another block of
memory, and since the API uses linkage pointers, such pointers would
become invalid.

This fix allocates the linkage array in advance, according to the
linkage_limit option.