Python bindings: Allow testing if Sentence.parse fails
Currently, there will just be no linkages in that case, but it not
possible to find out if an error occurred in the library.

Add bool test on the returned linkages iterator object (False indicates
that an error occurred).

Also add a comment on Sentence.split, indicating the meaning of its
returned value.