Python bindings: SAT bug fix
The SAT parser returned NULL when there are no more linkages.
However, currently this is not checked. The result is an assert() crash
when trying to fetch more linkages than exist.

Fix this by adding a bool check on the Linkage object (False iff it is
NULL), and a check for that in sentence_parse.next().