tests.py: Make it able to be invoked from any directory
Currently, if not invoked from its own directory, it uses an
environment variable "srcdir" in order to find its directory, from
which it reads its data files.

Make the use of "srcdir" redundant. However, if it is supplied
its value is still used for the location of the data files.

After this change, tests.py can be invoked from the root of the source
directory without a need to set "srcdir".