tests.py: Make the tests independent of equal-cost linkage order
See the problem description in PR #353.

The implemented algo is a combination of (2) and (3):

The is implemented as described - by the linkages graphical
representation: The graphical representation of equal-cost linkages are
fetched, sorted, and returned one by one.  A group of equal-sorted
linkages is so handled only when needed, e.g. if only one linkage is
tested, only the first group of equal-cost linkages is handled.

Also fix the data files accordingly.

Implementation notes:
The parse class of the Sentence class is "decorated" by a filter that
sorts the equal-cost linkages. In this way tests.py doesn't need any
other change. The new code is in a new local module lg_testutils.py.
Putting it in a separate module also has the benefit that it can
easily be used in other program. A program that mimics link-parser
(yet unpublished) uses it in order to make it easy to provide the
correct linkages for tests.py test comparisons.