Allow calling chart_parse() with and w/o nulls
... i.e. when min_null_count=0 and max_null_count=sent->length.
This fixes issue #303.

- Current state:
When chart_parse() is called with min_null_count=0 and max_null_count>0,
it sometimes malfunctions and gives less linkages than it should. This
is because power_prune() (invoked from prepare_to_parse() which invokes
pp_and power_prune()) has an optimization for null_count=0, which should
not be used for null_count>0 (because it removes connectors that can be
used then). It seems this is not documented.

This means that library users currently need to implement a 2-step
parsing if they want to find linkages with null_count>0 for sentences
with no complete linkage (but they don't know this, at least initially).
If they neglect to do that, they will find that link-parser sometimes
mysteriously gives parses that their program doesn't.

- This change:
Power-pruning is moved from prepare_to_parse() to be inside
chart_parse() in its loop of parsing with an increased null count, so
it is invoked (at most) twice - for null-count=0 and null-count>0.
Because power_prune() changes the disjuncts, if there is a need to
reparse with null_count>0, the original sentence disjuncts are restored
before power-pruning is performed again.

Also: Allow calling with a large max_null_count. This is handy if the
sentence length is not known (because an explicit split has not been
done).

- Speed improvement:
If a linkages with minimal null-links (including 0) is to be produced
for a sentence, this fix has a significant speed improvement for
sentences that don't have a complete linkage, because a part of
prepare_to_parse(), which takes significant CPU time (especially for
Russian), needs to run only once (alloc_count_context() is done only
once too, but this doesn't save a significant CPU time.)

Note that no overhead is added if the parsing is done in the "old way",
i.e with two steps as in link-parser.

- Link parser:
The link-parser program currently doesn't take benefit of that, because
it uses a 2-step parsing, needed for batch processing and the optional
corpus step. It can be changed to use the new way for interactive usage
when the "corpus" option is not in effect.