SAT: Fixes a crash when no linkages are found
The crash is due to a variable I recently added, which is uninitialized.
Test sentence (have to be a sentence with no complete linkages):
about people attended
From some reason there is no problem when using it in link-parser.
However, when using it with the Python binding, a crash happens in
sat_free_linkages().