Test suite: Test "No complete linkages found" for SAT
It validates the following:
- No crashes ...
- That the recently add check in linkgrammar.py of the SAT sentinel
  linkage NULL works fine.

FIXME:
For some reason, this modification causes a memory leak reporting as
follows (the only LG library leak reporting from tests.py):

Direct leak of 120 byte(s) in 1 object(s) allocated from:
    #0 0x7f0bbf0c197a in malloc (/lib64/libasan.so.2+0x9897a)
    #1 0x7f0bb1f294ca in xalloc /usr/local/src/link-grammar-devel/satfixes/link-grammar/utilities.c:418
    #2 0x7f0bb1e80b38 in parse_options_create /usr/local/src/link-grammar-devel/satfixes/link-grammar/api.c:102
    #3 0x7f0bb2514f4a in _wrap_parse_options_create ../../bindings/python/lg_python_wrap.cc:3622
    #4 0x7f0bbe02e795 in PyEval_EvalFrameEx (/lib64/libpython2.7.so.1.0+0xe5795)

Indirect leak of 64 byte(s) in 1 object(s) allocated from:
    #0 0x7f0bbf0c197a in malloc (/lib64/libasan.so.2+0x9897a)
    #1 0x7f0bb1f294ca in xalloc /usr/local/src/link-grammar-devel/satfixes/link-grammar/utilities.c:418
    #2 0x7f0bb1f0c0ba in resources_create /usr/local/src/link-grammar-devel/satfixes/link-grammar/resources.c:62
    #3 0x7f0bb1e812af in parse_options_create /usr/local/src/link-grammar-devel/satfixes/link-grammar/api.c:145
    #4 0x7f0bb2514f4a in _wrap_parse_options_create ../../bindings/python/lg_python_wrap.cc:3622
    #5 0x7f0bbe02e795 in PyEval_EvalFrameEx (/lib64/libpython2.7.so.1.0+0xe5795)

As far as I could validate, it happens due to the default ParseOption()
in the following added line:

    def parse_sent(self, text, po=ParseOptions()):

Adding print() in the __del__ method of class ParseOptions shows that
for some reason it is not called on exit of DBasicParsingTestCase (when
everything in it should get out of scope). It is called as needed in
other cases.