link-parser: Remove setting of disjunct_cost
This is a value that is tied to the design of dictionaries, and its
default is set in the library as needed. Because it is set also here,
when it gets changed in the library, there is a need to make the same
change here.

Also fix a bug due to my move of the command-line flags handling -
one cannot override disjunct_cost by the cost-max command-line flag.