Fix buffer overflow on #include of an empty file
When reading an empty file, fgets returns NULL without writing to the
buffer. So the buffer contains garbage.

The get_file_contents() function then just returns a pointer to the
buffer, and the caller tries to analyze this garbage. The result
depends on actual garbage.

Fix it by assigning '\0' to the first character of the buffer.