Port sat-solver to use external minisat 2.2 library
This fix does 2 things:
1. --enable-sat-parser (default setup) validates that minisat2 exists.
XXX It doesn't check that it is version >= 2.2, even though this is the
required version, because I don't know how to do that (it doesn't seem
to have a version information. Maybe a function that didn't exist before
2.2 can be checked). However, it seems that minisat packages are
available for install on all/most systesms are only from version 2.2.

2. Make the needed code changes to adapt the code from minisat 2.0 to 2.2.
Especially, clear the include hell that caused the wrong assert() to be
defined in the code.
Note:
printStats() is not available in 2.2. The old copy can be just
included in the sat-solver code instead (the varables it prints are
still available). Not done for now.