configure SAT: Adapt for certain minisat2.2 source variation
For some reason, there are several variations of minisat 2.2 source
code. Ubuntu and Fedora come with different Solver::setPriority.
Add a check for that, and also for changing Lit() to mkLIT() a
variation seen elsewhere).

Also check for the existence of minisat include directory, taking care
to allowing its override.