configure SAT: Use system or bundled library
The SAT-solver is still enabled by default.
If configure finds a minisat library + headers in the system, it uses
it. Minisat 2.2 is required, but the version is not explicitly checked.
However, the checked include structure is of version 2. In case there
is no system library and/or no includes, the bundled system is
automatically used.

In addition, --enable-sat-solver=bundled can be used to force using the
bundled minisat library.