New autoconf macros for save/restore variables
An autoconf macro can creates variables that prevents using it again
in the same configure file.

These change introduce macros that solve the problem by allowing to
discard the new variables that are created by such a macro.  If some of
these variables have to persist, they just need to be created in a head
(usually with empty value).

XXX A more sophisticated sed program is needed in order to handle edge
cases of multi-line variable values, but the provided one is enough
for the current use case.