Fix a bad bug that apparently doesn't have any impact
The intention of the offending code line is to compares the lengths of
the uppercase part of two connectors. If they are not the same, it
means the connectors don't match.

However, there is a bug and a variable is compared to itself (gcc 6.2.0
detected that, while 5.3.1 didn't). This means that a non-equal
uppercase connectors could leak to the rest of the code, which has a
comparison assuming they have an equal length.

However, it turned out that in practice they always have an equal
length. It seems this is because the hashing is very good and it just
happens that connectors with different uppercase length have never the
same hash value.