Configure minisat: Require zlib for minisat
If zlib-devel/zlib1g-dev is not installed, the compilation fails
because zlib.h is not found (system minisat cannot even detected in
that case).

So require it for a more graceful failure, with a message (by the
AX_CHECK_ZLIB):

configure: error: either specify a valid zlib installation with
 --with-zlib=DIR or disable zlib usage with --without-zlib

The bundled minisat doesn't include te files that actually use zlib,
so linking with -lz is not implemented.

(FIXME? Maybe "zlib" can be just #define'ed to "stdio"? But the various
packages will still require zlib because it is used in included headers.)