--enable_sat-solver: Link the link-grammar solib with libstdc++
(This fix doesn't apply to OS X.)
This is needed in order that it will be stand-alone regarding C++ symbol
resolution, instead of via the minisat solib.

See issue #428.

FIXME? This change also adds libstdc++ as a dependency of the
link-parser binary, even so it doesn't needed it.  This can be fixed if
needed with some more efforts.