autogen.sh: Workaround a Cygwin m4 include problem
Without this fix, m4/varcheckpoint.m4 is not loaded on Cygwin.

It is not clear to me what causes this problem (maybe some mismatched or
not up to date autotools).  This fix is restricted to Cygwin only
because it is harmful on OS X (due to apparent bug there of creating a
shell script "install-sh" without execute permission and then trying to
invoke it directly).