Cygwin: Add CXX flag _DEFAULT_SOURCE for "new" (200809) functions
Cygwin 2.6.0 (just released) has locale_t (previous versions didn't
have).  This definition is needed in ordered that locale_t changes will
be properly exposed in the header files.

The sat-solver source code doesn't use it, but has #include "utilities.h",
which uses it and bombs if it is not defined properly.