corpus.h: Put api-types.h/utilities.h behind the include guard
In addition, change the dummy functions so they still give the same
results.

Motivation:
Since the corpus code is not in use for now but corpus.h is still
included in various files, problems in these 2 include files just
clutter the output with numerous warnings (especially on Windows, in
which they still generate some minor warnings). They also introduce
unneeded dependencies for the source files which include corpus.h
(directly or indirectly).