configure sat-solver: if no system minisat lib, don't check for headers
If the library is not found, don't check for minisat headers.
This is a cosmetic change only, as MINISAT_INCLUDES (the system headers)
is not used anyway (no AC_SUBST) if the bundled library is used.

Note that if the headers are not found, MINISAT_LIBS is not set to
nothing, since it is also not used if the bundled library is used
(again, no AC_SUBST for it in that case).