Fix info message when a spell checker cannot be enabled
This message is produce when a spell checker is configured but there is
no spell dictionary of the particular language.

Fixes:
1. Indicate the language, because if multiple dictionaries are used it
is not clear to which language the message applies.
2. Print the message only on verbosity>0 (also less clutter in the test
suite output).

Also remove the TODO comments (quoted below, with explanation of removal):
 - Set the spell option to 0, to signify no spell checking is done.
It cannot be done in the LG library. It is more appropriate for an
application (like link-parser) to do it, but there is currently no API
to check the status of the spell checker for a given dict.
 - On verbosity >= 1, add a detailed message on the reason. */
In the speller setup function there is already a message in case of a
system error in setting the speller.

This message should be moved to spellcheck-*.c when the hard-coded
language mapping is removed in removed and the per-dict locale is used
instead.