Windows: Add LG_DLLPATH User Macro
It is for external libraries DLLs needed for using the parser binary and
the bindings. Currently it should point to the directory of "regex.dll".
(If libedit is add, its DLL should be added too, with a ";" delimiter.)

How it is used:
It is incorporated (by "post-build.bat") into link-parser.bat.
Also, "make-check.py" takes it from "Local.props" in run-time.