configure.ac: Add -D__STDC_FORMAT_MACROS and -D__STDC_LIMIT_MACROS
This fixes a failure in sat-solver compilation, due to undefined
macros PRI* and INT32_* in the minisat header file Options.h (both in
the bundled and unbundled minisat versions).

The added __STDC_* definitions are needed for CentOS-7 in order to expose
the C99 format and limit macros also in C++ source code.

Add them always in case there are other distributions that define them.