tests.py: Use setUpClass() for speedup
Currently, several test cases have setUP(), which is a function that
gets invoked before every test in the test case. Since these setUP()
functions include an expansive operation Directory() that returns always
the same result, doing it repetitively instead of reusing its result
adds a vast overhead.

Change it to setUpClass() instead, which is a function that is invoked
once before all the tests in a test case.

As before, changes of self.something in an individual test are private
to that test.

This change results in a 50+% speedup.