tests.py: Delete class variables on end of each test case
As a result of the previous change, running the tests with memory check
reports on huge leaks because the dictionary and parse-options stuff
are not getting freed since they are class variables and the classes
are still referenced.

In order to allow a clean run (so real memory leaks can be detected) just
delete the class variables at the end of each test case that defines
them.