Python ParseOptions(): Disallow positional arguments
Currently something like ParseOptions(2,3) is accepted, and jut sets the
options that happen to be in this argument positions.

This fix emulates the Python3 bare "*" feature that disallows
positional arguments, so it works in Python 2 too.

Note that this fix also validates (unintentionally, due to the way it works)
that the option name is correct. If it is not correct, the following
exception happens ('...' is the invalid option):
TypeError: __init__() got an unexpected keyword argument '...'"

Also note that this is detected also by pylint(!) so IDEs flag an error
on this sentence while editing:
po = ParseOptions(invalid=1)