link-parser: Fix using stale stack memory for string arguments
Fix using stale memory for setting the -debug= and -test= command lines
options.

The problem can be seen on certain systems only (MacOs, Fedora 25,
Ubuntu 16.04) but not on others (among them Fedora 23, up-to-date Gentoo,
Windows 10).

I first detected the problem after I upgraded from Fedora 23 to 25.
On it, setting debug or test from the command line leaves them usually
empty, and sometimes garbled. The problem also appears on MacOS.

It is interesting that this problem was not detected by AddressSanitizer
and Valgrind (including on Fedora 25, in which the problem is clearly
seen).

This fix allocates memory for the command line arguments on the heap.

The code indentation was not fixed in this commit so the commit
diff remains clear.