word_start_another_alternative: Fix partial debug messages
The new error facility allows to print a message in part and still have
it printed as one complete message.

However, this gets broken if another message is interleaved.

Fix it for one case in which this happens (maybe there are others).
A better fix is to use append_string() in this particular case, but this
may be considered an overkill.