Reissue and mark w/"WS_INDICT" spell result if it is in the dict
This fixes two problems:

1. Currently when starting to process a wordgraph word which is a
spell-guess, it is not reissued.  However, if this word gets
alternatives (due to a split or lowercasing) it is forgotten because it
is not initially reissued as one of the alternatives. This is bad if it
is also a dict word.  For example, if the result of a spell correction
is "I", then it will get an alternative "i", and the original "I" (which
is in the dict) will be forgotten.

2. Spell results are currently marked w/"WS_INDICT" unconditionally.
This is bad if the word is actually *not* in the dict.
For example, "ii'm" get corrected to "I'm", which is not in the dict.
(The ability to split spell results were added in the wordgraph
release.)

How it is fixed:
Currently non-spell-guessed words are issued and marked w/"WS_INDICT"
if they are in the dict.
This fix does it for every word (including spell-guess ones).