possible_connection() w/null_count=0: A word gap can now collapse
If a gap between words contains consecutive "optional" words and they
are missing in the linkage, the words may need a direct link.

Fix the pruning optimization for non-adjacent when preparing for a parse
with null-count=0, since they may become adjacent in such a case.

The rest of changes here are for providing the sentence pointer for
this check.