do_count(): Account for "optional" words
The trick is to allow them to remain out of the linkage (i.e. be "null
words" - words with no links) while not incrementing the null_count.

Also here:
1. flatten_wordgraph() doesn't issue any more an empty-word when
a slot is optional. Instead, it just marks it as optional.
2. remove_empty_words() now needs to remove the optional words
which happen to be "null words".

FIXME?
Rename remove_empty_words() to remove_optional_null_words().
Rename empty_word_encountered to optional_word_encountered.