Adapt print_expression_parens() for the SAT-parser
The SAT parser naively joins all X_node expressions using "or", and
this check used to give an error due to that, preventing a convenient
debugging.
Just accept it and continue to show the implied expression (but mark it
with '!').

Example: Word slot for "это":
это.=       LLDTS+					# pruned by expression_prune()
это.msi     (Wd- & MVAi+) or S+
это.p       EI+
эт.=        LLFFU+ or LLDVV+

Original print_expression_parens():
(EI+) or (Wd- & MVAi+) or S+
ERROR! Unexpected list!

The full expression as shown by print_exp_mem() (see the <--- mark):
e=0x603001d60710: OR_type cost=0.000000
  E_list=0x6020014fb1d0 (0x603001d47150 0x603001d47270 0x603001d47090)
  e=0x603001d47150: AND_type cost=0.000000
    E_list=0x6020014d4650 (0x603001d47120)
    e=0x603001d47120: CONNECTOR_type cost=0.000000
    con=EI dir=+ multi=0
  e=0x603001d47270: OR_type cost=0.000000
    E_list=0x6020014d46d0 (0x603001d47210 0x603001d47240)
    e=0x603001d47210: AND_type cost=0.000000
      E_list=0x6020014d4690 (0x603001d471b0 0x603001d471e0)
      e=0x603001d471b0: CONNECTOR_type cost=0.000000
      con=Wd dir=- multi=0
      e=0x603001d471e0: CONNECTOR_type cost=0.000000
      con=MVAi dir=+ multi=0
    e=0x603001d47240: CONNECTOR_type cost=0.000000
    con=S dir=+ multi=0
  e=0x603001d47090: OR_type cost=0.000000	    <--- problematic "or" here
    E_list=0x6020014d4630 (0x603001d47030 0x603001d47060)
    e=0x603001d47030: CONNECTOR_type cost=0.000000
    con=LLFFU dir=+ multi=0
    e=0x603001d47060: CONNECTOR_type cost=0.000000
    con=LLDVV dir=+ multi=0

Using this fix:
(EI+) or (Wd- & MVAi+) or S+ or! (LLFFU+ or LLDVV+)